requirements.txt
# Core Dependencies
Flask==3.0.0
flask-socketio==5.3.6
eventlet==0.35.2

# Computer Vision & ML
opencv-python>=4.8.1.78
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.0.232
numpy>=1.24.3,<2.0
deep-sort-realtime>=1.3.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Scheduling
APScheduler==3.10.4

# Image Processing
imageio==2.33.0
Pillow==10.1.0

# Geometry Processing
shapely==2.0.2

# HTTP Requests
requests==2.31.0

# Timezone Handling
pytz==2024.1

# Configuration
python-dotenv==1.0.0

# Utilities
click==8.1.6
Jinja2==3.1.2
MarkupSafe==2.1.5
Werkzeug==3.0.1

# Security
cryptography==41.0.7
PyJWT==2.7.0

# Parsing
PyYAML==6.0.1
